[refactor] Moved check method
